Updated 1.0.0
Fixed Multiplayer not Loading properly or Moving - The 3D Internet plugin was experiencing an issue since we implemented the server check. First of all, the server check required the server ip to be passed to several processes using move.js functions, and second, the movedata was incorrectly reading the server id for approval. All is corrected and working again!