Updated with WalkTheWeb 3.7.3
Extreme Load Zone in Community - To assist in 3D internet when entering zones, we are now returning the Extreme Load action zone id as the extremeloadzoneid. This applies to connect/community.php and connect/communities.php.

Admin Check Connection Functions removed - In the 3D Internet admin channel, there was a loop interval for checking the connection like a heartbeat. This has been removed since the new code enhancements and socket.io update handles the connection and reconnect a lot better.

Movement Tracking Updates - In the 3D Internet plugin, the movement tracking has been updated to always connect (when movement is turned on in admin), but now it only tracks when you enter or exit an extreme load zone for the community using socket.io rooms.

Added Try Catch to Animation Loading - We added a try-catch around the functions that load new avatar animations. This allows us to suppress errors in loading if an animation is not found, while still loading the remaining animations.

Avatar Transitions with Smoke - The numerous transitions for avatars that use smoke to enter a scene were sometimes causing errors if the smoke was removed before the process completed. We added conditions to check for smoke before using it in the function.

Updated Actionzone Object - The actionzone object uses moldname but it was not in the original newActionZone function. It has been added.

3D Internet Plugin Massive Update - In the past couple months we have fully rebuilt the Multi-player server. Latest Linux version, NGinx, PHP, Node.js, socket.io, Discord, etc... Then we made big improvements to the connect and disconnect functionality for admin channels, movement, and chat. This project will go on to next month, but it is working much better so we decided to publish out the latest changes. While this update may only have a few posted updates, it has a complete rewrite on the back-end server. It is all for the better in the long run. More to come!

Catch Download Errors - The Download errors from file_get_contents were not being caught by the code. We added the @ and a test and now it catches the errors, keeping them from the users.

Updated Socket.io Client Files - We updated the socket.io client files to the latest version 4.7.2.

CORS transport Update - We added a transport reference pointing to websocket on the 3D Internet Plugin connections for the Admin, Move, and Chat functionality. This appears to have corrected the occasional CORS error.

getAddThingList Moved to 3D Internet Function - getAddThingList JavaScript function was moved to the 3D Internet Plugin under scripts/franchises.js. This includes all references to call it.

getAddBuildingList Moved to 3D Internet Function - getAddBuildingList JavaScript function was moved to the 3D Internet Plugin under scripts/franchises.js

Prep Check Date Function - The prepCheckDate function had the month and day reversed in the conditional output. This caused invalid date format in some cases. It has been corrected.

Add Building to Community Duplication - When adding a 3D Building to a 3D Community, the list of your 3D Buildings was being duplicated. The code was moved to the 3D Internet Plugin, but was not removed from the main admin menu code. It has been corrected.

Corrected Download Image Paths - We corrected the download image paths to point to the 3dnet.walktheweb.com server instead of the 3d.walktheweb.com server.

Width on Email Settings - The Email Settings form was too narrow for the content. We adjusted the width to match similar forms.

Added Missing Animations - The install was not including all of the necessary animations. We added the missing animations to the proper uploads/avatars/3b9bt5c70igtmqux/animations folder. This folder gets copied to all of the initial avatars animations folders during install. You can copy them manually if needed.